Move player from core to beatwave (#10)
