+++ src/beatwave/main.cpp
@@ -12,12 +12,13 @@
 
 #include <core/tunnel.hpp>
 #include <core/captures.hpp>
-#include <core/player.hpp>
 #include <core/lineartransition.hpp>
 #include <core/lineartransitionbuilder.hpp>
 #include <core/seqcombinator.hpp>
 #include <core/util.hpp>
+
 #include <beatwave/game.hpp>
+#include <beatwave/player.hpp>
 
 const sf::Int32 SCREEN_WIDTH = 1280;
 const sf::Int32 SCREEN_HEIGHT = 720;
